+++ EssentialFeed/EssentialFeed/Feed Cache/LocalFeedLoader.swift
@@ -35,11 +35,3 @@ public final class LocalFeedLoader {
     }
 }
 
-public protocol FeedStore {
-    typealias DeletionCompletion = (Error?) -> Void
-    typealias InsertionCompletion = (Error?) -> Void
-    func deleteCacheFeed(completion: @escaping DeletionCompletion)
-    func insert(_ items: [FeedItem], timestamp: Date, completion: @escaping InsertionCompletion)
-}
-
-
